+++ db/migrate/20190311135819_add_include_medications_to_exports.rb
@@ -1,0 +1,5 @@
+class AddIncludeMedicationsToExports < ActiveRecord::Migration[6.0]
+  def change
+    add_column :exports, :include_medications, :boolean, null: false, default: false
+  end
+end