+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     - Dates where only the year and month are known: Jan 1990
   - Project editors can configure additional variables to be collected alongside
     medication name and start and stop dates
+  - Medications can be exported to CSV
 
 ### Refactoring
 - Remove skipping_utf8 in search forms as Rails 6 no longer enforces utf8